backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
cryptography==47.0.0
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.3.1
faiss-cpu==1.8.0
fastapi==0.110.0
filelock==3.19.1
fsspec==2025.10.0
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
groq==1.0.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.13
Jinja2==3.1.6
joblib==1.5.3
lxml==6.1.0
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.3.1
packaging==26.2
pillow==11.3.0
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.17.0
PyMuPDF==1.23.22
PyMuPDFb==1.23.22
pyparsing==3.3.2
python-docx==1.0.1
python-dotenv==1.0.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==2.5.1
sniffio==1.3.1
starlette==0.36.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.8.0
tqdm==4.67.3
transformers==4.57.6
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.27.1
